+++ src/de/jrk/neuralNetwork/activationFunctions/ActivationFunction.java
@@ -3,5 +3,5 @@ package de.jrk.neuralNetwork.activationFunctions;
 public abstract class ActivationFunction {
 	public final static TangensHyperbolicus tangensHyperbolicus = new TangensHyperbolicus();
 	
-	public abstract double function(double x);
+	public abstract float function(float x);
 }
